+++ common/model/src/main/java/module-info.java
@@ -22,8 +22,8 @@
  * <li>{@link is.codion.common.model.UserPreferences}
  * <li>{@link is.codion.common.model.CancelException}
  * <li>{@link is.codion.common.model.FilterModel}
- * <li>{@link is.codion.common.model.selection.SingleItemSelection}
- * <li>{@link is.codion.common.model.selection.MultiItemSelection}
+ * <li>{@link is.codion.common.model.selection.SingleSelection}
+ * <li>{@link is.codion.common.model.selection.MultiSelection}
  * <li>{@link is.codion.common.model.condition.ConditionModel}
  * <li>{@link is.codion.common.model.condition.TableConditionModel}
  * <li>{@link is.codion.common.model.summary.SummaryModel}